test: fix and split test_server_init_session_visibility.py into 4 focused files
Splits the 1675-line test_server_init_session_visibility.py into 3 focused
files plus a split guard, aligning each new file with the source module it
tests. Also fixes two bugs uncovered during the split.

Bug fixes:
- F-C1-09: Remove contradictory mock_create_bg_task.assert_not_called()
  in test_skill_auto_gate_boot_no_quota_refresh_loop — the label-filtered
  assertion already expresses the correct intent.
- F-C1-10: Remove duplicate RetryReason.CLONE_CONTAMINATION dict key in
  test_routing_completeness.py — line 35 silently overwrote the correct
  evidence-keyword entry on line 41.

Split:
- test_session_type_visibility.py — TestSessionTypeVisibility +
  TestFeatureGateVisibility (server/_session_type.py)
- test_lifespan_fleet_boot.py — TestFleetAutoGateBoot +
  TestFleetAutoGateBootProjectDir + TestFoodTruckAutoGateBoot
  (server/_lifespan.py)
- test_lifespan_skill_boot.py — TestSkillAutoGateBoot +
  TestSkillAutoGateBootRegistry (server/_lifespan.py)
- test_server_init_session_visibility_split.py — structural guard

Registry updates:
- _FLEET_CLASS_MARKERS in tests/arch/test_feature_markers.py: key now
  points to the new test_session_type_visibility.py
- .autoskillit/test-source-map.json: 43 source-file entries updated to
  reference the new test files
