+++ tests/cli/test_routing_completeness.py
@@ -32,7 +32,6 @@ _EXPECTED_ROUTES: dict[RetryReason, tuple[str, str | None]] = {
     RetryReason.COMPLETED_NO_FLUSH: ("on_context_limit", None),
     RetryReason.EMPTY_OUTPUT: ("on_failure", None),
     RetryReason.PATH_CONTAMINATION: ("on_failure", None),
-    RetryReason.CLONE_CONTAMINATION: ("on_failure", None),
     RetryReason.THINKING_STALL: ("on_context_limit", "lifespan_started"),
     RetryReason.IDLE_STALL: ("on_context_limit", "lifespan_started"),
     RetryReason.EARLY_STOP: ("on_context_limit", "has_progress_evidence"),
